+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -4,42 +4,13 @@
 
 import CoreData
 
-public final class CoreDataFeedStore: FeedStore {
+public final class CoreDataFeedStore {
 	private let container: NSPersistentContainer
 	private let context: NSManagedObjectContext
 
 	public init(storeURL: URL, bundle: Bundle = .main) throws {
 		container = try NSPersistentContainer.load(modelName: "FeedStore", url: storeURL, in: bundle)
 		context = container.newBackgroundContext()
-	}
-
-	public func retrieve(completion: @escaping RetrievalCompletion) {
-		perform { context in
-			completion(Result {
-				try ManagedCache.find(in: context).map {
-					CachedFeed(feed: $0.localFeed, timestamp: $0.timestamp)
-				}
-			})
-		}
-	}
-	
-	public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-		perform { context in
-			completion(Result {
-				let managedCache = try ManagedCache.newUniqueInstance(in: context)
-				managedCache.timestamp = timestamp
-				managedCache.feed = ManagedFeedImage.images(from: feed, in: context)
-				try context.save()
-			})
-		}
-	}
-
-	public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-		perform { context in
-			completion(Result {
-				try ManagedCache.find(in: context).map(context.delete).map(context.save)
-			})
-		}
 	}
 
 	func perform(_ action: @escaping (NSManagedObjectContext) -> Void) {
